# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tonatiuh Trading SaaS API
  version: 1.0.0
  description: Multi-tenant API for accounts, subscriptions, exchange connections, and trading bots.
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/organizations:
    get:
      tags: [Organizations]
      operationId: listOrganizations
      responses: {'200': {description: User workspaces and current organization}, default: {$ref: '#/components/responses/Error'}}
...
  /api/v1/admin/stats:
    get:
      tags: [Platform Admin]
      operationId: adminStats
      responses: {'200': {description: Platform counters}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/users:
    get:
      tags: [Platform Admin]
      operationId: adminListUsers
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}, {$ref: '#/components/parameters/Search'}]
      responses: {'200': {description: Users without credentials}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/users/{id}/status:
    patch:
      tags: [Platform Admin]
      operationId: adminChangeUserStatus
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      requestBody: {$ref: '#/components/requestBodies/PlatformStatus'}
      responses: {'200': {description: User status changed}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/users/{id}/revoke-sessions:
    post:
      tags: [Platform Admin]
      operationId: adminRevokeUserSessions
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      responses: {'200': {description: Refresh sessions revoked}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/organizations:
    get:
      tags: [Platform Admin]
      operationId: adminListOrganizations
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}, {$ref: '#/components/parameters/Search'}]
      responses: {'200': {description: Organizations and aggregate counts}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/organizations/{id}/status:
    patch:
      tags: [Platform Admin]
      operationId: adminChangeOrganizationStatus
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      requestBody: {$ref: '#/components/requestBodies/PlatformStatus'}
      responses: {'200': {description: Organization status changed}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/audit-events:
    get:
      tags: [Platform Admin]
      operationId: adminListAuditEvents
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses: {'200': {description: Platform audit events}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/data-subject-requests:
    get:
      tags: [Platform Admin, Compliance]
      operationId: adminListDataSubjectRequests
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}, {name: status, in: query, schema: {type: string, enum: [REQUESTED, IN_PROGRESS, COMPLETED, REJECTED]}}]
      responses: {'200': {description: Data-subject request queue and overdue state}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/data-subject-requests/{id}:
    patch:
      tags: [Platform Admin, Compliance]
      operationId: adminUpdateDataSubjectRequest
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, additionalProperties: false, required: [status], properties: {status: {type: string, enum: [IN_PROGRESS, COMPLETED, REJECTED]}, rejectionReason: {type: string, minLength: 1, maxLength: 1000}}}}}
      responses: {'200': {description: Data-subject request updated and audited}, '409': {description: Request is already final}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/feature-flags:
    get:
      tags: [Platform Admin]
      operationId: adminListFeatureFlags
      responses: {'200': {description: Feature flag configuration and override counts}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/feature-flags/{key}:
    patch:
      tags: [Platform Admin]
      operationId: adminUpdateFeatureFlag
      parameters: [{$ref: '#/components/parameters/FeatureFlagKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, additionalProperties: false, required: [enabled, rolloutPercentage, expectedVersion, changeReason], properties: {enabled: {type: boolean}, rolloutPercentage: {type: integer, minimum: 0, maximum: 100}, expectedVersion: {type: integer, minimum: 1}, changeReason: {type: string, minLength: 1, maxLength: 500}}}}}
      responses: {'200': {description: Feature flag updated}, '409': {description: Optimistic locking conflict}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/feature-flags/{key}/organizations/{organizationId}:
    put:
      tags: [Platform Admin]
      operationId: adminSetFeatureFlagOverride
      parameters: [{$ref: '#/components/parameters/FeatureFlagKey'}, {$ref: '#/components/parameters/OrganizationId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, additionalProperties: false, required: [enabled, expectedVersion, changeReason], properties: {enabled: {type: boolean}, expectedVersion: {type: integer, minimum: 1}, changeReason: {type: string, minLength: 1, maxLength: 500}}}}}
      responses: {'200': {description: Organization override set}, '409': {description: Optimistic locking conflict}, default: {$ref: '#/components/responses/Error'}}
    delete:
      tags: [Platform Admin]
      operationId: adminRemoveFeatureFlagOverride
      parameters: [{$ref: '#/components/parameters/FeatureFlagKey'}, {$ref: '#/components/parameters/OrganizationId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, additionalProperties: false, required: [expectedVersion, changeReason], properties: {expectedVersion: {type: integer, minimum: 1}, changeReason: {type: string, minLength: 1, maxLength: 500}}}}}
      responses: {'200': {description: Organization override removed}, '409': {description: Optimistic locking conflict}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/feature-flags/{key}/organizations:
    get:
      tags: [Platform Admin]
      operationId: adminListFeatureFlagOverrides
      parameters: [{$ref: '#/components/parameters/FeatureFlagKey'}, {$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses: {'200': {description: Organization overrides for a feature flag}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/payments:
    get:
      tags: [Platform Admin]
      operationId: adminListPayments
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses: {'200': {description: Sanitized payment records}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/payments/{id}/refund:
    post:
      tags: [Platform Admin]
      operationId: adminRefundPayment
      parameters: [{$ref: '#/components/parameters/ResourceId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, additionalProperties: false, required: [reason], properties: {reason: {type: string, minLength: 1, maxLength: 500}}}}}
      responses: {'201': {description: Full refund succeeded}, '202': {description: Refund is pending reconciliation}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/refunds:
    get:
      tags: [Platform Admin]
      operationId: adminListRefunds
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses: {'200': {description: Refund records without provider snapshots}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/system:
    get:
      tags: [Platform Admin]
      operationId: adminSystemStatus
      responses: {'200': {description: Worker and queue aggregates}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/email/dead-letters:
    get:
      tags: [Platform Admin]
      operationId: adminListEmailDeadLetters
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses: {'200': {description: Dead letters with masked recipients}, default: {$ref: '#/components/responses/Error'}}
  /api/v1/admin/email/dead-letters/{id}/retry:
    post:
      tags: [Platform Admin]
      operationId: adminRetryEmailDeadLetter
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      responses: {'202': {description: Dead letter queued for retry}, default: {$ref: '#/components/responses/Error'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    metricsToken: {type: http, scheme: bearer, bearerFormat: opaque metrics token}
    emailWebhookToken: {type: http, scheme: bearer, bearerFormat: opaque provider webhook token}
  parameters:
    ResourceId: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
    OrganizationId: {name: organizationId, in: path, required: true, schema: {type: string, format: uuid}}
    FeatureFlagKey: {name: key, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_]{2,63}$'}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 128}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
    Search: {name: search, in: query, schema: {type: string, maxLength: 320}}
  responses:
    Error:
      description: Standard API error
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  requestBodies:
...
    PlatformStatus:
      required: true
      content: {application/json: {schema: {type: object,additionalProperties: false,required: [status],properties: {status: {type: string,enum: [ACTIVE,SUSPENDED]}}}}}
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties: {code: {type: string}, message: {type: string}, requestId: {type: string}, details: {}}
...